layers: Create dedicated shader interface file
